% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint}
\alias{blueprint}
\title{Creates a character matrix which specifies which effects to estimate and
which effects to constrain to a non-zero value}
\usage{
blueprint(effects, use)
}
\arguments{
\item{effects}{A data frame that contains information on which effects to
estimate or constrain to a value other than zero. Each row
represents one effect. The \code{effects} object  must have three
columns: column 1 has the variable names for the predictors,
column 2 has the variable names for the outcomes, and column
3 has the names of the effects (if the effect should be estimated)
or the numeric value the effect should be set at (if the effect
is not to be estimated but rather to be set to a value other
than zero).}

\item{use}{A vector with the variable names that will be used in the
stability-informed model}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A character matrix
}\if{html}{\out{</div>}}
}
\description{
Creates a character matrix which specifies which effects to estimate and
which effects to constrain to a non-zero value
}
\examples{
\dontrun{
effects <- data.frame(predictor = "X_0", outcome = "Y", name = "CLxy")
use <- c("X", "Y")
CreateBlueprint(effects, use)
}

}
\keyword{internal}
