% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SIModel}
\alias{print.SIModel}
\title{stim print function}
\usage{
\method{print}{SIModel}(out, digits = 2)
}
\arguments{
\item{out}{A stim object}

\item{digits}{Number of digits to round to}
}
\value{
A result table
}
\description{
stim print function
}
\keyword{internal}
