% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stim.R
\name{print.stim}
\alias{print.stim}
\title{stim print function}
\usage{
\method{print}{stim}(out)
}
\arguments{
\item{out}{A stim object}
}
\value{
Overview of model
}
\description{
stim print function
}
\keyword{internal}
