% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultTable.R
\name{resultTable}
\alias{resultTable}
\title{Creating a Result Table from a stim object}
\usage{
resultTable(modelList)
}
\arguments{
\item{modelList}{A list of SIModel inputs and outputs}
}
\value{
A result table
}
\description{
Creating a Result Table from a stim object
}
\keyword{internal}
