% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.R
\name{symbMatrix}
\alias{symbMatrix}
\title{Create a symbolic covariance matrix}
\usage{
symbMatrix(blueprint, residualcov)
}
\arguments{
\item{blueprint}{A character matrix which specifies which effects to estimate and
which effects to constrain to a non-zero value}

\item{residualcov}{A list with both the lavaan syntax for the residual covariance
and a dataframe with the variable names}
}
\value{
A list of character matrices: A symbolic covariance matrix and a symbolic
psi matrix
}
\description{
Create a symbolic covariance matrix
}
\keyword{internal}
