% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic.R
\name{\%*\%}
\alias{\%*\%}
\alias{symbMultiplication}
\title{Internal function used to do symbolic matrix multiplication using the Ryacas
R package}
\usage{
symbMultiplication(x1, x2)
}
\arguments{
\item{x1}{A Ryacas object.}

\item{x2}{A Ryacas object.}
}
\description{
Internal function used to do symbolic matrix multiplication using the Ryacas
R package
}
\keyword{internal}
