% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{blueprint}
\alias{blueprint}
\title{Creates a character matrix which specifies which effects to estimate and
which effects to constrain to a non-zero value}
\usage{
blueprint(effects, use)
}
\arguments{
\item{effects}{A data frame that contains information on which cross-lagged effects to
estimate or constrain to a value other than zero. Each row
represents one effect. The \code{effects} object  must have four
columns: column 1 has the variable names for the predictors,
column 2 has the variable names for the outcomes, column
3 has the names of the estimated effect or the value an unestimated
effect should be constrained to, and column 4 has information
on whether the effect should be estimated. This table can be
created by using the \code{effectTable} function}

\item{use}{A vector with the variable names that will be used in the
stability-informed model}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A character matrix
}\if{html}{\out{</div>}}
}
\description{
Creates a character matrix which specifies which effects to estimate and
which effects to constrain to a non-zero value
}
\keyword{internal}
