% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectTable.R
\name{effectTable}
\alias{effectTable}
\title{Create a parameter table}
\usage{
effectTable(model)
}
\arguments{
\item{model}{An object with the model description for the cross-sectional model in lavaan syntax}
}
\value{
A data frame with information on which effects to estimate and which to constrain. Each
row represents one effect and specifies which variable is the predictor and outcome of the effect.
The name column contains information on whether the effect should be estimated--if the name is a
character (e.g., CLxy)--or if the effect should be constrained to a value other than 0 (e.g., .3)
}
\description{
Create a parameter table
}
\examples{

#estimate effect from X to Y
#constrain effect from Y to X to .3
model <- c('Y ~  X
            X ~ .3 * Y')

effectTable(model)
}
