% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanEq.R
\name{lavaanEq}
\alias{lavaanEq}
\title{Create lavaan syntax based on the blueprint matrix}
\usage{
lavaanEq(blueprint, S)
}
\arguments{
\item{blueprint}{A character matrix which specifies which effects to estimate and
which effects to constrain to a non-zero value}

\item{S}{Sample covariance matrix}
}
\value{
A character vector
}
\description{
Create lavaan syntax based on the blueprint matrix
}
\examples{
\dontrun{
S <- matrix( c(1, .3, .3, 1), nrow = 2, ncol = 2)
colnames(S) <- rownames(S) <- c("X", "Y")

blueprint <- matrix( c("ARx", "CLxy", "ARy", "0"), nrow = 2, ncol = 2 )
colnames(blueprint) <- c("X", "Y")
rownames(blueprint) <- c("X_0", "Y_0")

lavaanEq(blueprint, S)
}
}
\keyword{internal}
