% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelImpliedEq.R
\name{modelImpliedEq}
\alias{modelImpliedEq}
\title{Get the model implied symbolic equations for the auto-regressive effects and
the covariances between the phantom variables}
\usage{
modelImpliedEq(S, blueprint, stability, residualcov)
}
\arguments{
\item{S}{Sample covariance matrix}

\item{blueprint}{A character matrix which specifies which effects to estimate and
which effects to constrain to a non-zero value}

\item{stability}{A data frame that contains stability information for each
variable in the model. If unnamed SIM will assume the stability
values are in the same order as the provided data set/
covariance matrix.}

\item{residualcov}{A list with both the lavaan syntax for the residual covariance
and a dataframe with the variable names}
}
\value{
A character vector with the model implied equations for the autoregressive
effects and the phantom covariances.
}
\description{
Get the model implied symbolic equations for the auto-regressive effects and
the covariances between the phantom variables
}
\examples{

S <- matrix( c(1, .3, .3, 1), nrow = 2, ncol = 2)
colnames(S) <- rownames(S) <- c("X", "Y")

blueprint <- matrix( c("ARx", "CLxy", "ARy", "0"), nrow = 2, ncol = 2 )
colnames(blueprint) <- c("X", "Y")
rownames(blueprint) <- c("X_0", "Y_0")

 stability <- data.frame(X = .3, Y = .3)

 residualcov <- list(Syntax = 'X ~~ RCovXY * Y',
                     Variables = data.frame(V1 = "X", V2 = "Y", Name = "RCovXY"))

 modelImpliedEq(S, blueprint, stability, residualcov)

}
