% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stim.R
\name{summary.stim}
\alias{summary.stim}
\title{Summary method for \code{stim} objects}
\usage{
\method{summary}{stim}(object, ...)
}
\arguments{
\item{object}{An object of class \code{stim}}

\item{...}{Not used}
}
\value{
A print out containing the results for a set of Stability Informed Models
}
\description{
Summarize a set of Stability Informed Models
}
\examples{
\donttest{
S <- matrix(.3, 5, 5)
diag(S) <- 1
set.seed(69)
example_data <- as.data.frame(MASS::mvrnorm(n = 300, mu = rep(0, 5), Sigma = S,))

model <- 'X5 ~ X4 + X3'
stability <- c(X3 = .3, X4 = .2, X5 = .1)

modelFit <- stim(data = example_data, model = model, stability = stability)

summary(modelFit)
}

}
\seealso{
\code{\link{stim}}
}
